// packages/common/src/sites/_internal/types.ts
export enum SearchCategories {
  DATA = "components.search.category_tabs.data",
  SITES = "components.search.category_tabs.sites",
  DOCUMENTS = "components.search.category_tabs.documents",
  APPS_AND_MAPS = "components.search.category_tabs.apps_and_maps",

  // The following entries are currently used, but will be phased out
  EVENTS = "components.search.category_tabs.events",
  INITIATIVES = "components.search.category_tabs.initiatives",
}